backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0

